% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeCoverage.R
\name{aggregate_coverage}
\alias{aggregate_coverage}
\title{aggregate the coverage vectors across every window in the tfbs file}
\usage{
aggregate_coverage(bed_file, tfbs, window)
}
\arguments{
\item{coverage}{a dictonary of chromosome names and coverage vectors}

\item{tfbs:}{path to the tfbs file}

\item{window:}{the size of the window to aggregate over}
}
\value{
: aggregated coverage vector
}
\description{
aggregate the coverage vectors across every window in the tfbs file
}
