% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeCoverage.R
\name{aggregate_coverage_over_tfbs}
\alias{aggregate_coverage_over_tfbs}
\title{Return the tfbs from the csv inputted file}
\usage{
aggregate_coverage_over_tfbs(TFBS_csv, sample_bed, offset = 500)
}
\arguments{
\item{csome_num}{the chromosome number of interest string like "chr1"}
}
\description{
Return the tfbs from the csv inputted file
}
