% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeRatio.R
\name{get_controles_di}
\alias{get_controles_di}
\title{Helper function: Detecting dinucleosome lengths of controls}
\usage{
get_controles_di(control_frag_sizes)
}
\arguments{
\item{control_frag_sizes}{A dataframe for the a control data
(healthy individual) cfDNA fragment lengths.
This data frame has 1 column, showing the cfDNA fragment lengths for
one control individual.}
}
\value{
Return the control length dataframe with only dinucleosome sizes
as defined in the research by Katsman et.al (1).
}
\description{
Helper function: Detecting dinucleosome lengths of controls
}
\references{
1- Katsman, E., Orlanski, S., Martignano, F., Fox-Fisher, I., Shemer,
R., Dor, Y., Zick, A., Eden, A., Petrini, I., Conticello, S. G.,
Berman, B. P. (2022). Detecting cell-of-origin and cancer-specific
methylation features of cell-free DNA from Nanopore sequencing.
Genome biology, 23(1), 158.
https://doi.org.myaccess.library.utoronto.ca/10.1186/s13059-022-02710-1
}
