% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeCoverage.R
\name{get_coverage}
\alias{get_coverage}
\title{Nucleosome coverage plots for chromosome positions. Used to find Transcription
Factor Binding Sites (TFBS)}
\usage{
get_coverage(sample_bed, chr, start, end)
}
\arguments{
\item{sample_bed}{A dataframe for the a sample data
(unknown if cancerous or not individual)
with at least 3 columns: first column is a string representing the chromosome
number for the cfDNA for instance: "chr1",
the second column are integer values indicating the start location of
the read and third column are integer values indicating end location of
the read.
These are the patient who we want to check if they have cancer or not.
The fragment lengths
of patients will be compared to these healthy data similar to what is done in
the paper by Katsman et. al (1).}

\item{chr}{the chromosome of interest. Is a string of format for example
"chr1" for chromosome 1}

\item{start}{the position of the cfDNA fragment}

\item{end}{the ending position of the cfDNA fragment

#TODO:
add references to content above.}
}
\value{
Return an dataframe of nucleosome coverage:
value refer to the number of cfDNA fragments that cover the region of the
nucleosome.
}
\description{
TFBS can be used to subtype cancer DNA. Transcription Factors (TFs)bind to
specific regions of specific kinds of DNA, to allow for their transcription.
The binding of the TFs allow for generation of RNA from that DNA region.
It is important to note that regions that are binded by TFs are not bound
by nucleosomes (to allow for binding of the TF protein). However, other
regions of the DNA are bound by these nucleosomes, protecting the DNA from
degradation. cfDNA samples that do not have TF bound, do not have
nucleosomes bound at the region where TF usually binds. Hence, they are
not protected from nuclease degradation. This lead to the cfDNA to be cut at
the TF binding sites since those sites are unprotected. Hence, cfDNA fragment
ends if many of the cfDNA are cut at that certain region can be used as an
indicator for TF binding sites. Hence, coverage plots are used to detect the
amount of coverage by the cfDNA molecules on the DNA. If the coverage is low,
that could indicate the existance of a TFBS in that region.
The coverage plots are created using the nucleosomeCoverage data structure.
}
\details{
Note 1: The data to generate examples are from the illumina website (4).
This example data is not representative of real patient data due to
limitations to access of real patient data.
However, users can use their own real patient datasets for
analysis.
More examples can be found at https://github.com/Yasamin-Nourijelyani/CfDNAfragmentomics/tree/master/inst/extdata

Note 2:
BED files can be generate from BAM files using bedtools bamtobed function
in the command line (3).

Since BED files are commonly
generated by computational biologists and are easily convertible, from
BAM files, they can be used as input. Otherwise, tab seperated
.txt files that generate data in the same format of bed files
with the same columns as described in the parameters are also
excepted.
}
\examples{
\dontrun{


sample_bed_file <- system.file("extdata", "d1.bed", package = "CfDNAfragmentomics")
sample_bed <- read.table(sample_bed_file, header = FALSE, sep="\t",stringsAsFactors=FALSE, quote="")

## basic usage

# Example 1:

## Note: example does not work because the chromosome1 length is shoerter
than the chromosome length in sample bed file

cov <- nucleosomeCoverage(sample_bed = sample_bed, 237995947)
cov
# check for instance
cov[2985822, 1]

}


}
\references{
1- Katsman, E., Orlanski, S., Martignano, F., Fox-Fisher, I., Shemer,
R., Dor, Y., Zick, A.,
Eden, A., Petrini, I., Conticello, S. G., & Berman, B. P. (2022).
Detecting cell-of-origin and cancer-specific methylation features of
cell-free DNA from Nanopore sequencing.
Genome biology, 23(1), 158.
https://doi.org.myaccess.library.utoronto.ca/10.1186/s13059-022-02710-1

2- “Mann Whitney U Test in R Programming.” GeeksforGeeks, 28 Dec. 2021,
https://www.geeksforgeeks.org/mann-whitney-u-test-in-r-programming/.

3- Quinlan, Aaron R., Kindlon, Neil. "bamtobed" readthedocs, May 26, 2022,
https://bedtools.readthedocs.io/en/latest/content/tools/bamtobed.html

4- Illumina, https://support.illumina.com/downloads/nextera-flex-for-enrichment-BED-files.html

5- “The T-Test.” JMP,
https://www.jmp.com/en_ca/statistics-knowledge-portal/t-test.html#:~
:text=t\%2DTest\%20assumptions&text=The\%20data\%20are\%20continuous.,
The\%20distribution\%20is\%20approximately\%20normal.
}
