% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucleosomeCoveragePython.R
\name{nucleosomeCoverage1}
\alias{nucleosomeCoverage1}
\title{Plot coverage from pythpn file
sample_bed_file <- system.file("extdata", "p1.bed", package = "CfDNAfragmentomics")
sample_bed <- read.table(sample_bed_file, header = FALSE, sep="\t",stringsAsFactors=FALSE, quote="")}
\usage{
nucleosomeCoverage1(bed_file, tfbs, window, name)
}
\description{
library(readr)
TFBS_loc <- read_csv("inst/extdata/TFBS_loc.csv")
}
\details{
pip install tqdm
}
