% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeCoveragePlot.R
\name{nucleosomeCoveragePlot}
\alias{nucleosomeCoveragePlot}
\title{Plot of the coverage for cfDNA fragments}
\usage{
nucleosomeCoveragePlot(coverage_df)
}
\arguments{
\item{sample_bed}{A dataframe for the a sample data
(unknown if cancerous or not individual)
with at least 3 columns: first column is a string representing the chromosome
number for the cfDNA for instance: "chr1",
the second column are integer values indicating the start location of
the read and third column are integer values indicating end location of
the read.
These are the patient who we want to check if they have cancer or not.
The fragment lengths
of patients will be compared to these healthy data similar to what is done in
the paper by Katsman et. al (1).
These samples are compared to the controls using a t-test.}

\item{chromosome_length}{The integer length of the chromosome being analysed
Lengths can also be referenced from http://www.insilicase.com/Web/Chromlen.aspx
given the chromosome number.}
}
\value{
Returns a plot of cfDNA coverage
}
\description{
A visualization function that generates a coverage plot showing
nucleosome nucleosome patient data to visually
detect TFBS.
}
\examples{
\dontrun{

# Example 1:

sample_bed_file <- system.file("extdata", "d1.bed", package = "CfDNAfragmentomics")
sample_bed <- read.table(sample_bed_file, header = FALSE, sep="\t",stringsAsFactors=FALSE, quote="")

## basic usage

library(ggplot2)
plot <- nucleosomeCoveragePlot(sample_bed = sample_bed, 237995947)




}


}
\references{
1- Katsman, E., Orlanski, S., Martignano, F., Fox-Fisher, I., Shemer,
R., Dor, Y., Zick, A.,
Eden, A., Petrini, I., Conticello, S. G., & Berman, B. P. (2022).
Detecting cell-of-origin and cancer-specific methylation features of
cell-free DNA from Nanopore sequencing.
Genome biology, 23(1), 158.
https://doi.org.myaccess.library.utoronto.ca/10.1186/s13059-022-02710-1

2- Illumina, https://support.illumina.com/downloads/nextera-flex-for-enrichment-BED-files.html

3- “The T-Test.” JMP,
https://www.jmp.com/en_ca/statistics-knowledge-portal/t-test.html#:~
:text=t\%2DTest\%20assumptions&text=The\%20data\%20are\%20continuous.,
The\%20distribution\%20is\%20approximately\%20normal.
}
