% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeRatio.R
\name{nucleosomeRatio}
\alias{nucleosomeRatio}
\title{Wilcox-Test of nucleosome ratios for patient and control cfDNA fragment
reads data.}
\usage{
nucleosomeRatio(
  controls_bed,
  sample_bed,
  di_nucleosome_p_value = 0.05,
  mono_nucleosome_p_value = 0.05
)
}
\arguments{
\item{controls_bed}{A dataframe for the a control data (healthy individual)
with at least 3 columns: first column is a string representing the chromosome
number for the cfDNA for instance: "chr1",
the second column includes integer values indicating the start location of
the read and third column includes integer values indicating end location of
the read.
These are the healthy individuals who do not have cancer.The fragment lengths
of patients will be compared to these healthy data similar to what is done in
the paper by Katsman et. al (1).}

\item{sample_bed}{A dataframe for the a sample data
(unknown if cancerous or not individual)
with at least 3 columns: first column is a string representing the chromosome
number for the cfDNA for instance: "chr1",
the second column are integer values indicating the start location of
the read and third column are integer values indicating end location of
the read.
These are the patient who we want to check if they have cancer or not.
The fragment lengths
of patients will be compared to these healthy data similar to what is done in
the paper by Katsman et. al (1).
These sample data are compared to the controls using a t-test.}

\item{di_nucleosome_p_value}{A numeric value between 0 and 1
indicating the p-value for the dinuclesome ratio that is considered
significant. Default value is 0.05.}

\item{mono_nucleosome_p_value}{A numeric value between 0 and 1
indicating the p-value for the dinuclesome ratio that is considered
significant. Default value is 0.05}
}
\value{
Return a list of outputs:
The function returns a list of outputs.
This list contains a real number: ttest_mono_pvalue, a real number: ttest_di_pvalue,
and a boolean value: cancerous.
ttest_mono_pvalue is the p-value calculated for the difference in mononucleosome
lengths between patient and controls, ttest_di_pvalue is the p-value calculated
for the difference in dinucleosome lengths, and if the patient input files shows significant shorter
mononucleosome and dinucleosome sizes compared to the control with respect to the
inputted p-values that are considered significant, cancerous will be TRUE.
and cancerous will be FALSE otherwise. Note that even if the p-values show significance,
if the patient cfDNA lengths are not shorter, then the patient is not considered as being cancerous.
}
\description{
Fragmentation sizes of cfDNA (cell free DNA) molecules are potential
cancer biomarkers (1).
Hence, to find if a patient data file contains cancer information,
it can be compared with a control dataset with known healthy cfDNA fragments.
The Wilcox test is a non-parametric test used to determine
significance of the difference in fragment
size of the control and patient cfDNA lengths to determine if the patient
data contains the cancer biomarker.
Here, we are comparing two independent samples, which are the
control and patient data, and we are making inference about the state
(cancer positive or negative) of the
population of cfDNA molecules in the patient as being potentially cancer
positive or negative. Both control and patient
data inputted to the functions are assumed to be reads for the specific loci
corresponding to the cancer type of interest. Hence, both the control
and patient data are required in the analysis to ensure the loci
being analysed are corresponding to the specific cancer type of interest.
Additionally, performing a non-parametric wilcox-test does not assume that the data is
normally distributed (5).
It is also assumed that the data is real human data and contains
both mono-nucleosome and di-nucleosome length data.
}
\details{
Note 1: The data to generate examples are from the illumina website (4).
This example data is not representative of real patient data due to
limitations to access of real patient data.
However, users can use their own real patient datasets for
analysis.
More examples can be found at https://github.com/Yasamin-Nourijelyani/CfDNAfragmentomics/tree/master/inst/extdata

Note 2:
BED files can be generate from BAM files using bedtools bamtobed function
in the command line (3).

Since BED files are commonly
generated by computational biologists and are easily convertible, from
BAM files, they can be used as input. Otherwise, tab seperated
.txt files that generate data in the same format of bed files
with the same columns as described in the parameters are also
excepted.
}
\examples{
\dontrun{

controls_bed_file <- system.file("extdata", "p1.bed", package = "CfDNAfragmentomics")
controls_bed <- read.table(controls_bed_file, header = FALSE, sep="\t",stringsAsFactors=FALSE, quote="")

sample_bed_file <- system.file("extdata", "d1.bed", package = "CfDNAfragmentomics")
sample_bed <- read.table(sample_bed_file, header = FALSE, sep="\t",stringsAsFactors=FALSE, quote="")

## basic usage

# Example 1:

ratio <- nucleosomeRatio(controls_bed = controls_bed, sample_bed = sample_bed, 0.3, 0.3)
ratio$cancerous

# Example 2:

ratio <- nucleosomeRatio(controls_bed = controls_bed, sample_bed = sample_bed)
ratio$cancerous

}


}
\references{
1- Katsman, E., Orlanski, S., Martignano, F., Fox-Fisher, I., Shemer,
R., Dor, Y., Zick, A.,
Eden, A., Petrini, I., Conticello, S. G., & Berman, B. P. (2022).
Detecting cell-of-origin and cancer-specific methylation features of
cell-free DNA from Nanopore sequencing.
Genome biology, 23(1), 158.
https://doi.org.myaccess.library.utoronto.ca/10.1186/s13059-022-02710-1

2- “Mann Whitney U Test in R Programming.” GeeksforGeeks, 28 Dec. 2021,
https://www.geeksforgeeks.org/mann-whitney-u-test-in-r-programming/.

3- Quinlan, Aaron R., Kindlon, Neil. "bamtobed" readthedocs, May 26, 2022,
https://bedtools.readthedocs.io/en/latest/content/tools/bamtobed.html

4- Illumina, https://support.illumina.com/downloads/nextera-flex-for-enrichment-BED-files.html

5- “The T-Test.” JMP,
https://www.jmp.com/en_ca/statistics-knowledge-portal/t-test.html#:~
:text=t\%2DTest\%20assumptions&text=The\%20data\%20are\%20continuous.,
The\%20distribution\%20is\%20approximately\%20normal.
}
