% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyCfDNAfragmentomics.R
\name{runShinyCfDNAfragmentomics}
\alias{runShinyCfDNAfragmentomics}
\title{Use the shiny CfDNAfragmentomics}
\usage{
runShinyCfDNAfragmentomics()
}
\value{
No return value but open up a Shiny page.
}
\description{
Launch Shiny App for CfDNAfragmentomics
}
\details{
A function that launches the Shiny app for CfDNAfragmentomics
The purpose of this app is only to run the Shiny
app. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
}

}
\references{
1- Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}

2- Silva, A. (2022) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.
}
