% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomeCoverage.R
\name{get_coverage}
\alias{get_coverage}
\title{Get the coverage at every locus in the interval \link{start, end}}
\usage{
get_coverage(bed_file)
}
\arguments{
\item{bed_file:}{The bed file to get the coverage from
A dataframe for the a sample data
(unknown if cancerous or not individual)
with at least 3 columns: first column is a string representing the chromosome
number for the cfDNA for instance: "chr1",
the second column are integer values indicating the start location of
the read and third column are integer values indicating end location of
the read.
These are the patient who we want to check if they have cancer or not.
The fragment lengths
of patients will be compared to these healthy data similar to what is done in
the paper by Katsman et. al (1).

#TODO:
add references to content above.}
}
\value{
A list of coverage values
}
\description{
Note 1: The data to generate examples are from the illumina website (4).
This example data is not representative of real patient data due to
limitations to access of real patient data.
However, users can use their own real patient datasets for
analysis.
More examples can be found at https://github.com/Yasamin-Nourijelyani/CfDNAfragmentomics/tree/master/inst/extdata
}
\details{
Note 2:
BED files can be generate from BAM files using bedtools bamtobed function
in the command line (3).

Since BED files are commonly
generated by computational biologists and are easily convertible, from
BAM files, they can be used as input. Otherwise, tab seperated
.txt files that generate data in the same format of bed files
with the same columns as described in the parameters are also
excepted.
}
\examples{
\dontrun{


sample_bed_file <- system.file("extdata", "d1.bed", package = "CfDNAfragmentomics")
sample_bed <- read.table(sample_bed_file, header = FALSE, sep="\t",stringsAsFactors=FALSE, quote="")

## basic usage

# Example 1:

## Note: example does not work because the chromosome1 length is shoerter
than the chromosome length in sample bed file

cov <- get_coverage(bed_file = sample_bed)
cov
# check for instance
cov[2985822, 1]

}



}
\references{
1- Katsman, E., Orlanski, S., Martignano, F., Fox-Fisher, I., Shemer,
R., Dor, Y., Zick, A.,
Eden, A., Petrini, I., Conticello, S. G., & Berman, B. P. (2022).
Detecting cell-of-origin and cancer-specific methylation features of
cell-free DNA from Nanopore sequencing.
Genome biology, 23(1), 158.
https://doi.org.myaccess.library.utoronto.ca/10.1186/s13059-022-02710-1

2- “Mann Whitney U Test in R Programming.” GeeksforGeeks, 28 Dec. 2021,
https://www.geeksforgeeks.org/mann-whitney-u-test-in-r-programming/.

3- Quinlan, Aaron R., Kindlon, Neil. "bamtobed" readthedocs, May 26, 2022,
https://bedtools.readthedocs.io/en/latest/content/tools/bamtobed.html

4- Illumina, https://support.illumina.com/downloads/nextera-flex-for-enrichment-BED-files.html

5- “The T-Test.” JMP,
https://www.jmp.com/en_ca/statistics-knowledge-portal/t-test.html#:~
:text=t\%2DTest\%20assumptions&text=The\%20data\%20are\%20continuous.,
The\%20distribution\%20is\%20approximately\%20normal.
}
